# backend/internal/delivery/http/openapi.yaml
openapi: 3.0.3
info:
  title: Platform RT API
  description: |
    OpenAPI 3.0 specification for the Sitransparan RT/RW backend.

    Autentikasi: `Authorization: Bearer <jwt>` (HS256, valid 24 jam).
    Role: `superadmin`, `admin_rt`, `resident`.
    Konvensi akses pada deskripsi endpoint:
    - PUBLIC: tanpa autentikasi
    - AUTH: semua role terautentikasi
    - ADMIN: superadmin + admin_rt
    - SUPERADMIN: superadmin saja

    Tenant context hanya diambil dari JWT claims yang terverifikasi (header
    X-Tenant-ID, query param, dan subdomain TIDAK dipercaya). Endpoint publik
    `/api/v1/t/{slug}/...` me-resolve tenant dari slug di path.
  version: 2.0.0
servers:
  - url: http://localhost:8081
    description: Local server (Port 8081)
  - url: /
    description: Current host relative path
paths:
...
  /api/v1/residents:
    get:
      summary: List warga - NIK tidak dikembalikan plaintext (ADMIN)
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer } }
        - { name: offset, in: query, schema: { type: integer } }
      responses:
        '200':
          description: '{ data: Resident[], total, limit, offset }'
    post:
      summary: Buat warga - NIK dienkripsi AES-256-GCM (ADMIN)
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResidentInput'
      responses:
        '201':
          description: Warga dibuat
  /api/v1/residents/{id}:
    get:
      summary: Detail warga (ADMIN)
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Resident
        '404':
          description: Tidak ditemukan (termasuk resource tenant lain)
    put:
      summary: Update warga (ADMIN)
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResidentInput'
      responses:
        '200':
          description: Warga diupdate
    delete:
      summary: Hapus warga (ADMIN)
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Warga dihapus
...
  /api/v1/residents/upload:
    post:
      summary: Upload file KTP/KK/dokumen (AUTH)
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
                type: { type: string, default: document }
      responses:
        '200':
          description: '{ file_url, type }'

  # ------------------------------------------------------------- Financial
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ResidentInput:
      type: object
      properties:
        full_name: { type: string }
        nik: { type: string, description: Dienkripsi AES-256-GCM di server }
        nik_hash: { type: string }
        kk_number: { type: string }
        gender: { type: string }
        birth_place: { type: string }
        birth_date: { type: string, format: date }
        address: { type: string }
        rt_rw: { type: string }
        phone: { type: string }
        is_head_of_family: { type: boolean }
        status: { type: string, enum: [pending, approved, rejected] }
        ktp_url: { type: string }
        kk_url: { type: string }